python_microservices/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
torch==2.7.1
torchvision==0.22.1
transformers==4.53.0
pillow==11.2.1
numpy==2.0.2
requests==2.32.4
python-multipart==0.0.20 